SALSA: Fix datatype for number of points parameter in segmentation functions
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@12866 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 89bd6cc9fac1d86f69bafa35d51bda4226a8972b

+++ application/SaLsA/de.uniol.inf.is.odysseus.classification/src/de/uniol/inf/is/odysseus/classification/segmentation/IEPFSegmentation.java
@@ -26,7 +26,7 @@ public class IEPFSegmentation extends AbstractFunction<Geometry> {
     private static final long           serialVersionUID = 4973220087210280849L;
 
     public static final SDFDatatype[][] accTypes         = new SDFDatatype[][] {
-            { SDFSpatialDatatype.SPATIAL_MULTI_POINT }, { SDFDatatype.DOUBLE }, { SDFDatatype.DOUBLE } };
+            { SDFSpatialDatatype.SPATIAL_MULTI_POINT }, { SDFDatatype.DOUBLE }, { SDFDatatype.INTEGER } };
     private final GeometryFactory       geometryFactory  = new GeometryFactory();
 
     @Override
